[WFCORE-5490] Move the resolver extension interfaces to the org.jboss.as.controller.extension package
